refactor(web): break-down-notable-persons-component
